refactor(solana): move encoding sources into :libs:solana:encoding
Move Instruction, Message, MessageHeader, LegacyMessage, VersionedMessage,
SolanaTransaction, ShortVec, MessageAddressLookupTable, and
AddressLookupTable out of :services:opencode and into the new module's
commonMain, keeping their packages. ShortVec stays internal, so
ShortVecTest.kt moves with it — internal visibility is module-scoped and
the test can no longer compile against :services:opencode once the type
lives elsewhere.

Extract SolanaTransaction.diff() into SolanaTransactionDiff.kt, staying in
:services:opencode: it depends on Differ.kt's printDiff/printMatch, which
use timber.log.Timber and can't move. Its MessageHeader.description
reference is inlined rather than exposed, since that property is internal
inside the new module and diff() was its only caller.

Replace the two javaClass-based equals() checks (Instruction, MessageHeader)
with `other !is X`, matching the KMP-safe pattern already used elsewhere
(Key32, PublicKey) — javaClass doesn't resolve on Kotlin/Native. Neither
class has subclasses, so this preserves current equals() semantics exactly.
