+++ libs/solana/encoding/src/commonMain/kotlin/com/getcode/opencode/internal/solana/ShortVec.kt
