+++ libs/solana/encoding/src/commonMain/kotlin/com/getcode/opencode/solana/Instruction.kt
@@ -29,7 +29,7 @@ data class Instruction(
 
     override fun equals(other: Any?): Boolean {
         if (this === other) return true
-        if (javaClass != other?.javaClass) return false
+        if (other !is Instruction) return false
 
         other as Instruction
