+++ libs/solana/encoding/src/commonMain/kotlin/com/getcode/opencode/solana/LegacyMessage.kt
