+++ libs/solana/encoding/src/commonMain/kotlin/com/getcode/opencode/solana/MessageHeader.kt
@@ -18,7 +18,7 @@ open class MessageHeader(
 
     override fun equals(other: Any?): Boolean {
         if (this === other) return true
-        if (javaClass != other?.javaClass) return false
+        if (other !is MessageHeader) return false
 
         other as MessageHeader
